requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0

drf-yasg==1.21.7

requests==2.32.3

psycopg2-binary==2.9.9
